Se crean los DTO y el manejador de excepciones y se añaden a las implementaciones de los servicios

+++ src/main/java/com/pruebaTecnica/spaceShips/service/implementation/SpaceShipImpl.java
@@ -1,10 +1,14 @@
 package com.pruebaTecnica.spaceShips.service.implementation;
 
+import com.pruebaTecnica.spaceShips.exceptions.SpaceShipException;
+import com.pruebaTecnica.spaceShips.mapper.SpaceShipDTOTOSpaceShip;
 import com.pruebaTecnica.spaceShips.persistence.entity.SpaceShip;
 import com.pruebaTecnica.spaceShips.persistence.repository.SpaceShipRepository;
+import com.pruebaTecnica.spaceShips.service.DTO.SpaceShipDTO;
 import com.pruebaTecnica.spaceShips.service.ISpaceShipService;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -14,9 +18,11 @@ import java.util.Optional;
 public class SpaceShipImpl implements ISpaceShipService {
 
     private final SpaceShipRepository spaceShipRepository;
+    private final SpaceShipDTOTOSpaceShip spaceShipDTOTOSpaceShip;
 
-    public SpaceShipImpl(SpaceShipRepository spaceShipRepository) {
+    public SpaceShipImpl(SpaceShipRepository spaceShipRepository, SpaceShipDTOTOSpaceShip spaceShipDTOTOSpaceShip) {
         this.spaceShipRepository = spaceShipRepository;
+        this.spaceShipDTOTOSpaceShip = spaceShipDTOTOSpaceShip;
     }
 
     @Override
@@ -25,28 +31,59 @@ public class SpaceShipImpl implements ISpaceShipService {
     }
 
     @Override
-    public Optional<SpaceShip> getSpaceShipById(Long id) {
-        return spaceShipRepository.findById(id);
+    public SpaceShip getSpaceShipById(Long id) {
+        //Se busca si existe la SpaceShip en BBDD
+        Optional<SpaceShip> optionalSpaceShip = spaceShipRepository.findById(id);
+        if(optionalSpaceShip.isEmpty()){
+            //Si no existe lanzo error
+            throw new SpaceShipException("Space ship not found", HttpStatus.NOT_FOUND);
+        }
+        //Si existe lo devuelvo
+        return optionalSpaceShip.get();
     }
 
     @Override
     public List<SpaceShip> getSpaceShipsByName(String name) {
-        return spaceShipRepository.findByNameContaining(name);
+        //Verifico si se genera un listado de SpaceShip con el nombre dado
+        List<SpaceShip> spaceShipList = spaceShipRepository.findByNameContaining(name);
+        if(spaceShipList.size() == 0){
+            //Si no se genera lanzo un error
+            throw new SpaceShipException("SpaceShips not found", HttpStatus.NOT_FOUND);
+        }
+        //Si genera la lista la devuelvo
+        return spaceShipList;
     }
 
     @Override
-    public SpaceShip createSpaceShip(SpaceShip spaceShip) {
+    public SpaceShip createSpaceShip(SpaceShipDTO spaceShipDTO) {
+        //Hago un mapeo para convertir el SpaceShipDTO en SpaceShip para poder persistir en la BBDD
+        SpaceShip spaceShip = spaceShipDTOTOSpaceShip.map(spaceShipDTO);
         return spaceShipRepository.save(spaceShip);
     }
 
     @Override
-    public SpaceShip updateSpaceShip(Long id, SpaceShip spaceShip) {
-        spaceShip.setId(id);
+    public SpaceShip updateSpaceShip(Long id, SpaceShipDTO spaceShipDTO) {
+        //Se busca si existe la SpaceShip en BBDD
+        Optional<SpaceShip> optionalSpaceShip = spaceShipRepository.findById(id);
+        if(optionalSpaceShip.isEmpty()){
+            //Si no existe lanzo error
+            throw new SpaceShipException("Space ship not found", HttpStatus.NOT_FOUND);
+        }
+        //Si existe actualizo
+        SpaceShip spaceShip = optionalSpaceShip.get();
+        spaceShip.setName(spaceShipDTO.getName());
         return spaceShipRepository.save(spaceShip);
     }
 
     @Override
     public void deleteSpaceShip(Long id) {
+        //Se busca si existe la SpaceShip en BBDD
+        Optional<SpaceShip> optionalSpaceShip = spaceShipRepository.findById(id);
+        if(optionalSpaceShip.isEmpty()){
+            //Si no existe lanzo error
+            throw new SpaceShipException("Space ship not found", HttpStatus.NOT_FOUND);
+        }
+        //Si existe lo elimino
         spaceShipRepository.deleteById(id);
     }
 }
